requirements.txt
tensorflow>=2.6
keras>=2.6
numpy>=1.19
pandas>=1.3
matplotlib>=3.4
seaborn>=0.11
opencv-python>=4.5
scikit-learn>=0.24
Pillow>=8.3
